--- a/mlir/test/Dialect/Rock/lowering_padding_kernel.mlir
+++ b/mlir/test/Dialect/Rock/lowering_padding_kernel.mlir
@@ -2,7 +2,7 @@
 // * The correct padding transformations are generated and added to the gemm
 
 // RUN: rocmlir-opt -rock-affix-params -rock-conv-to-gemm -rock-gemm-to-gridwise %s | FileCheck %s
-// CHECK-DAG: #[[$PAD_GEMMK:.*]] = #rock.transform_map{{.*}}Pad{0, 14} ["gemmKPad"] at [1] -> ["gemmK"] at [1]
+// CHECK-DAG: #[[$PAD_GEMMK:.*]] = #rock.transform_map{{.*}}Pad{0, 2} ["gemmKPad"] at [1] -> ["gemmK"] at [1]
 
 // CHECK-LABEL: func.func @rock_conv2d_kcyx_nchw_nkhw_padding_kernel
 // CHECK-SAME: %[[filter:.*]]: memref<32x128x2x3x3xf32>