This commit improves the init param ordering
to be sorted based on padding amount.
i.e. lowest padding is preffered.

When in equal padding cases, the original
ordering will be respected.

Doing so, exposed some params should be
dynamically filtered out according to
XdlopsCodeSelection.