@krzysz00 2022-09-09
Worth mentioning, you can set the base branch to be the other PR - that'll work fine