@manupak 2022-09-15
> Worth mentioning, you can set the base branch to be the other PR - that'll work fine

Other PR branch is in my fork right ? this is a missing feature request for GH for a while now :sigh.
https://github.com/orgs/community/discussions/4477#discussioncomment-1039627