@manupak 2022-10-05
Sure, How would you generally do that (other than running a random assortment)? Any pointers are appreciated