@krzysz00 2022-10-05
The way I usually do it is to run (after doing a `ninja check-rocmlir-build-only ci-performance-scripts`):

```
./bin/MIOpenDriver.py -b -o=some-name.csv
echo "description" >perf-run-date
[cd to a checkout of upstream]
ninja check-rocmlir-build-only ci-performance-scripts
./bin/MIOpenDriver.py -b -o=upstream.csv
echo "Upstream" >perf-run-date
[cd back]
python3 ./bin/perfRegressionReport.py [path-to-upstream].csv [path-to-here].csv output-file.html
```
then copy the file off, take a look at it, and post it here