@manupak 2022-10-06
https://gist.github.com/manupak/a1e679b89b1e05fe3db0a74b4879a266

Overall there is an increase there are few cases it drops..