@manupak 2022-10-06
ok some worse numbers does not seem to reproduce now and running multiple measurements to see...