@manupak 2022-10-06
I did three runs for the PR : 

1.) https://gist.github.com/manupak/887bcf5a3727f22ccfacff85b20603ac
2.) https://gist.github.com/manupak/1dcf61e381f472100f443bb2f8a52288
3.) https://gist.github.com/manupak/4156c2e8b9d09228d744a71ac1aee8b2

Generally looking at the changes only a single worse config stood out -- others more or less seemed like glitches.

| Direction | DataType | Layout | C |  H  |  W  |  K | Y | X | StrideH | StrideW | PaddingH | PaddingW | Unnamed: 0_new | MLIR TFlops (new) | Unnamed: 0_old | MLIR TFlops (old) | Current/Previous |
|:---------:|:--------:|:------:|:-:|:---:|:---:|:--:|:-:|:-:|:-------:|:-------:|:--------:|:--------:|:--------------:|:-----------------:|:--------------:|:-----------------:|:----------------:|
| fwd       | f16      | nhwc   | 3 | 230 | 230 | 64 | 7 | 7 |    2    |    2    |     0    |     0    |       68       |       11.75       |       68       |       21.11       |       0.56       |

It is interesting that it is also layout specific as the nchw layout seemed fine with config.